+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.0.3] - 2025-11-03
+
+### Added
+- Subscription status based monitoring flag.
+- Api token setup for registered user from dashboard.
+- Auto addition of protocol if only domain is added.
+- Custom craft headers to api calls.
+
+### Chore
+- Refactored the code base.
+- Fixed few url's in readme that were incorrect.
+
 ## [1.0.2] - 2025-10-09
 
 ### Fixed
@@ -61,7 +73,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Asset minification and optimization
 - Background processing for monitoring tasks
 
-[Unreleased]: https://github.com/Appfoster/upsnap/compare/v1.0.2...HEAD
-[1.0.2]: https://github.com/Appfoster/upsnap/releases/tag/v1.0.2
-[1.0.1]: https://github.com/Appfoster/upsnap/releases/tag/v1.0.1
-[1.0.0]: https://github.com/Appfoster/upsnap/releases/tag/v1.0.0
+[Unreleased]: https://github.com/Appfoster/upsnap/compare/1.0.3...HEAD
+[1.0.3]: https://github.com/Appfoster/upsnap/releases/tag/1.0.3
+[1.0.2]: https://github.com/Appfoster/upsnap/releases/tag/1.0.2
+[1.0.1]: https://github.com/Appfoster/upsnap/releases/tag/1.0.1
+[1.0.0]: https://github.com/Appfoster/upsnap/releases/tag/1.0.0